refactor: split AuTab0Viewer into separate components and files
- Created `CompactAccordion` and `ViewerOptionRow` for dense UI.
- Extracted `AuTab0OptionRow`, `AuTab0MapCategory`, and `AuTab0GeneralCategory` into their own files.
- Reduced vertical spacing and improved code organization.
- Adhered to state/stateless component placement rules.
